[6/n] make daft no-std compatible
Realized that there was nothing about daft that needed std or even alloc. Add
support for no-std and no-std-with-alloc builds.

I ended up splitting out all the implementations into separate modules gated by
their respective features.

Reviewers: andrewjstone

Reviewed By: andrewjstone

Pull Request: https://github.com/oxidecomputer/daft/pull/36
